# internal/iam/domain/tenant/controller.go
package tenant

import (
	"net/http"
	"tenant-crud-simply/internal/pkg/rest_err"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

// Controller interface define os métodos do controller de tenant
type Controller interface {
	Routes(routes gin.IRouter, authMiddleware gin.HandlerFunc)
	Create(c *gin.Context)
	Read(c *gin.Context)
	List(c *gin.Context)
	Update(c *gin.Context)
	Delete(c *gin.Context)
}

// controllerImpl implementa o Controller
type controllerImpl struct {
	service Service
}

// NewController cria uma nova instância do controller
func NewController(service Service) Controller {
	return &controllerImpl{
		service: service,
	}
}

// Routes registra as rotas do tenant
func (ctrl *controllerImpl) Routes(routes gin.IRouter, authMiddleware gin.HandlerFunc) {
	tenantGroup := routes.Group("/tenant")
	{
		tenantGroup.POST("/create", ctrl.Create)
		tenantGroup.POST("/create-blocked", authMiddleware, ctrl.Create)
		tenantGroup.GET("", ctrl.Read)
		tenantGroup.GET("/list", ctrl.List)
		tenantGroup.PATCH("/:uuid", ctrl.Update)
		tenantGroup.DELETE("", ctrl.Delete)
	}
}

// Create cria um novo tenant
// @Summary Cria um novo tenant
// @Description Cria um novo tenant no sistema
// @Tags Tenant
// @Accept json
// @Produce json
// @Param request body CreateTenantRequestDto true "Dados do tenant"
// @Success 201 {object} TenantResponseDto
// @Failure 400 {object} map[string]interface{}
// @Failure 500 {object} map[string]interface{}
// @Router /api/tenant/create-blocked [post]
func (ctrl *controllerImpl) Create(c *gin.Context) {
	var req CreateTenantRequestDto

	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error":   "invalid request",
			"details": err.Error(),
		})
		return
	}

	// Cria o modelo Tenant
	tenant := Tenant{
		Name:     req.Name,
		Document: req.Document,
		Live:     true,
		CreateAt: time.Now().UTC(),
		UpdateAt: time.Now().UTC(),
	}

	// Chama o serviço para criar
	created, err := ctrl.service.Create(c.Request.Context(), tenant)
	if err != nil {
		switch err {
		case ErrDocumentDuplicated:
			c.JSON(http.StatusConflict, gin.H{
				"error": "document already exists",
			})
		default:
			c.JSON(http.StatusInternalServerError, gin.H{
				"error":   "failed to create tenant",
				"details": err.Error(),
			})
		}
		return
	}

	c.JSON(http.StatusCreated, &TenantResponseDto{
		UUID:     created.UUID,
		Name:     created.Name,
		Document: created.Document,
		Live:     created.Live,
		CreateAt: created.CreateAt,
		UpdateAt: created.UpdateAt,
	})
}

// @Summary      Busca um Tenant
// @Description  Busca um tenant no sistema usando o UUID ou o Documento (CNPJ/CPF). Pelo menos um dos dois campos deve ser fornecido.
// @Tags         Tenant
// @Produce      json
// @Security     BearerAuth
//
// @Param        uuid query string false "UUID do tenant a ser buscado. (Ex: 8871abf3-ed11-4770-b986-e8d98d022d4f)"
// @Param        document query string false "Documento (CNPJ/CPF) do tenant a ser buscado. (Ex: 12345678901234)"
//
// @Success      200  {object}  TenantResponseDto  "Tenant encontrado com sucesso."
// @Failure      400  {object}  rest_err.RestErr    "Requisição inválida (UUID inválido, ou nenhum dos campos 'uuid'/'document' fornecido)."
// @Failure      404  {object}  rest_err.RestErr    "Tenant não encontrado com os dados fornecidos."
// @Failure      500  {object}  rest_err.RestErr    "Erro interno do servidor."
//
// @Router       /api/tenant [get]
func (ctrl *controllerImpl) Read(c *gin.Context) {
	var req ReadTenantRequestDto
	if err := c.ShouldBindQuery(&req); err != nil {
		restError := rest_err.NewBadRequestError("Parâmetros de busca inválidos.")
		c.JSON(restError.Code, restError)
		return
	}

	if req.UUID == "" && req.Document == "" {
		restError := rest_err.NewBadRequestError("É necessário fornecer o 'uuid' OU o 'document' para a busca.")
		c.JSON(restError.Code, restError)
		return
	}

	var tenantUUID uuid.UUID
	if req.UUID != "" {
		// Tentativa de conversão
		parsedUUID, err := uuid.Parse(req.UUID)
		if err != nil {
			// Captura erro de UUID mal formatado (retorna 400)
			restError := rest_err.NewBadRequestError("O UUID fornecido não é um formato válido.")
			c.JSON(restError.Code, restError)
			return
		}
		tenantUUID = parsedUUID
	}

	rTenant, err := ctrl.service.Read(c.Request.Context(), Tenant{
		UUID:     tenantUUID,
		Document: req.Document,
	})
	if err != nil {
		var restError *rest_err.RestErr
		switch err {
		case ErrNotFound:
			// Tratamento para 404
			restError = rest_err.NewNotFoundError(ErrNotFound.Error())
		case ErrInvalidInput:
			// Tratamento para 400 (Assumindo que InvalidInput no Read é uma falha na query/dados)
			restError = rest_err.NewBadRequestError(ErrInvalidInput.Error())
		default:
			// Tratamento para 500
			restError = rest_err.NewInternalServerError("Falha ao buscar tenant", nil)
		}

		c.JSON(restError.Code, restError)
		return
	}

	c.JSON(http.StatusOK, &TenantResponseDto{
		UUID:     rTenant.UUID,
		Name:     rTenant.Name,
		Document: rTenant.Document,
		Live:     rTenant.Live,
		CreateAt: rTenant.CreateAt,
		UpdateAt: rTenant.UpdateAt,
	})
}

// @Summary      Lista Tenants
// @Description  Retorna uma lista paginada de todos os tenants registrados no sistema.
// @Tags         Tenant
// @Produce      json
// @Security     BearerAuth
//
// @Param        page query int true "O número da página a ser retornada (deve ser >= 1)." default(1)
// @Param        pageSize query int true "O número de itens por página (máximo 100)." default(10)
//
// @Success      200  {object}  TenantsResponseDto  "Lista de tenants retornada com sucesso."
// @Failure      400  {object}  rest_err.RestErr    "Requisição inválida (parâmetros de paginação ausentes ou inválidos, ou pageSize > 100)."
// @Failure      500  {object}  rest_err.RestErr    "Erro interno do servidor."
//
// @Router       /api/tenant/list [get]
func (ctrl *controllerImpl) List(c *gin.Context) {
	var req ListTenantRequestDto
	if err := c.ShouldBindQuery(&req); err != nil {
		restError := rest_err.NewBadRequestError("Parâmetros de busca inválidos. Verifique 'page' e 'pageSize'.")
		c.JSON(restError.Code, restError)
		return
	}

	if req.PageSize > 100 {
		restError := rest_err.NewBadRequestError("É permitido um máximo de 100 listagens por página.")
		c.JSON(restError.Code, restError)
		return // Adicionado o 'return' para evitar execução posterior
	}

	lTenants, err := ctrl.service.List(c.Request.Context(), req.Page, req.PageSize)
	if err != nil {
		var restError *rest_err.RestErr
		restError = rest_err.NewInternalServerError("Falha ao buscar tenants", nil)
		c.JSON(restError.Code, restError)
		return
	}

	tenantResponses := make([]TenantResponseDto, len(lTenants))
	for i, t := range lTenants {
		tenantResponses[i] = TenantResponseDto{
			UUID:     t.UUID,
			Name:     t.Name,
			Document: t.Document,
			Live:     t.Live,
			CreateAt: t.CreateAt,
			UpdateAt: t.UpdateAt,
		}
	}
	c.JSON(http.StatusOK, &TenantsResponseDto{
		Tenants: tenantResponses,
		Page:    req.Page,
		Size:    req.PageSize,
	})
}

// @Summary      Atualiza um Tenant
// @Description  Atualiza dados de um tenant existente. O tenant a ser atualizado é identificado pelo UUID no path
// @Tags         Tenant
// @Accept       json
// @Produce      json
//
// @Param        uuid path string true "UUID do tenant a ser atualizado."
// @Param        request body UpdateTenantRequestDto true "Campos do tenant a serem atualizados. Apenas os campos presentes serão modificados."
//
// @Success      200  {object}  TenantResponseDto  "Tenant atualizado com sucesso."
// @Failure      400  {object}  rest_err.RestErr    "Requisição inválida (corpo JSON mal formatado, UUID inválido ou dados de entrada inválidos)."
// @Failure      404  {object}  rest_err.RestErr    "Tenant não encontrado para o UUID fornecido."
// @Failure      409  {object}  rest_err.RestErr    "Conflito (o novo 'document' fornecido já está em uso por outro tenant)."
// @Failure      500  {object}  rest_err.RestErr    "Erro interno do servidor."
//
// @Router       /api/tenant/{uuid} [patch]
func (ctrl *controllerImpl) Update(c *gin.Context) {
	uuidStr := c.Param("uuid")
	tenantUUID, err := uuid.Parse(uuidStr)
	if err != nil {
		restError := rest_err.NewBadRequestError("O UUID fornecido na URL não é um formato válido.")
		c.JSON(restError.Code, restError)
		return
	}

	var request UpdateTenantRequestDto
	if err := c.ShouldBindJSON(&request); err != nil {
		restError := rest_err.NewBadRequestError("Corpo JSON inválido ou mal formatado.")
		c.JSON(restError.Code, restError)
		return
	}

	uTenant := Tenant{
		UUID:     tenantUUID,
		Document: request.Document,
		Live:     *request.Live,
		Name:     request.Name,
		UpdateAt: time.Now().UTC(),
	}

	tenantUpdated, err := ctrl.service.Update(c.Request.Context(), &uTenant)

	if err != nil {
		var restError *rest_err.RestErr

		switch err {
		case ErrNotFound:
			restError = rest_err.NewNotFoundError(ErrNotFound.Error())
		case ErrDocumentDuplicated:
			restError = rest_err.NewConflictValidationError("O novo documento fornecido já está em uso por outro ", nil)
		case ErrInvalidInput:
			restError = rest_err.NewBadRequestError(ErrInvalidInput.Error())
		default:
			restError = rest_err.NewInternalServerError("Falha ao atualizar tenant", nil)
		}

		c.JSON(restError.Code, restError)
		return
	}

	c.JSON(http.StatusOK, &TenantResponseDto{
		UUID:     tenantUpdated.UUID,
		Name:     tenantUpdated.Name,
		Document: tenantUpdated.Document,
		Live:     tenantUpdated.Live,
		CreateAt: tenantUpdated.CreateAt,
		UpdateAt: tenantUpdated.UpdateAt,
	})
}

// @Summary      Deleta um Tenant
// @Description  Exclui permanentemente um tenant no sistema usando o UUID ou o Documento (CNPJ/CPF). Pelo menos um dos dois campos deve ser fornecido.
// @Tags         Tenant
// @Produce      json
//
// @Param        uuid query string false "UUID do tenant a ser excluído. (Ex: 8871abf3-ed11-4770-b986-e8d98d022d4f)"
// @Param        document query string false "Documento (CNPJ/CPF) do tenant a ser excluído. (Ex: 12345678901234)"
//
// @Success      204  {string} string "Tenant excluído com sucesso (No Content)."
// @Failure      400  {object}  rest_err.RestErr    "Requisição inválida (UUID inválido, ou nenhum dos campos 'uuid'/'document' fornecido)."
// @Failure      404  {object}  rest_err.RestErr    "Tenant não encontrado com os dados fornecidos."
// @Failure      500  {object}  rest_err.RestErr    "Erro interno do servidor."
//
// @Router       /api/tenant [delete]
func (ctrl *controllerImpl) Delete(c *gin.Context) {
	var req ReadTenantRequestDto
	if err := c.ShouldBindQuery(&req); err != nil {
		restError := rest_err.NewBadRequestError("Parâmetros de busca inválidos.")
		c.JSON(restError.Code, restError)
		return
	}

	if req.UUID == "" && req.Document == "" {
		restError := rest_err.NewBadRequestError("É necessário fornecer o 'uuid' OU o 'document' para a exclusão.")
		c.JSON(restError.Code, restError)
		return
	}

	var tenantUUID uuid.UUID
	if req.UUID != "" {
		parsedUUID, err := uuid.Parse(req.UUID)
		if err != nil {
			restError := rest_err.NewBadRequestError("O UUID fornecido não é um formato válido.")
			c.JSON(restError.Code, restError)
			return
		}
		tenantUUID = parsedUUID
	}

	err := ctrl.service.Delete(c.Request.Context(), Tenant{
		UUID:     tenantUUID,
		Document: req.Document,
	})

	if err != nil {
		var restError *rest_err.RestErr
		switch err {
		case ErrNotFound:
			restError = rest_err.NewNotFoundError(ErrNotFound.Error())
		case ErrInvalidInput:
			restError = rest_err.NewBadRequestError(ErrInvalidInput.Error())
		default:
			restError = rest_err.NewInternalServerError("Falha ao excluir tenant", nil)
		}

		c.JSON(restError.Code, restError)
		return
	}
	c.Status(http.StatusNoContent)
}